Fix lost chunks during streaming of large volumes of data.
The current streaming implementation allows for chunks of data
to be lost in VSSH's default select{} cases, if volumes of data
arrive at a rate too high to be processed.  The default cases
get selected when the write channel buffer becomes full and
would block.  This patch puts a client into a streaming mode
when the GetStream() function is called, which causes a blocking
select{} to be used until the streaming mode is ended.